+++ src/main/java/no/bachelor26/Tasks/DTO/TaskComponents.java
@@ -18,6 +18,6 @@ public class TaskComponents {
     private TaskData data;
     private String flag;
     private List<HintDTO> hints;
-    private Long unlocksTaskId;
+    private Integer unlocksTaskId;
 
 }
